+++ packages/third-ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @windrun-huaiin/third-ui
 
+## 23.1.1
+
+### Patch Changes
+
+- fix(mdx): local-md fix source build action, which can cause OOM
+  - support `portable-clerk`, `fumadocs-clerk`, `fumadocs-normal` 3 styles of TOC
+
 ## 23.1.0
 
 ### Minor Changes